conditionally render button on homepage and reorganize
