+++ client/src/App.css
@@ -1,8 +1,3 @@
-/* Homepage */
-.homepageContainer {
-  background-color: aqua;
-}
-
 /* CARD */
 .card-container {
   display: flex;
