+++ client/src/pages/Homepage/Homepage.css
@@ -1,0 +1,3 @@
+.homepageContainer {
+    background-color: aqua;
+}
