+++ client/src/pages/Marketplace/Marketplace.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faFire, faLeaf, faDroplet } from '@fortawesome/free-solid-svg-icons';
-import cardPhoto from '../images/Swinteger.webp';
-import cardPhoto2 from '../images/Leafarray.webp';
-import cardPhoto3 from '../images/Varchar.webp';
-import '../App.css';
+import cardPhoto from '../../images/Swinteger.webp';
+import cardPhoto2 from '../../images/Leafarray.webp';
+import cardPhoto3 from '../../images/Varchar.webp';
+import '../../App.css';
 
 const creatures = [
     {
